+++ App/xpmKit/Sources/Loader/GraphNode.swift
@@ -2,7 +2,7 @@ import Basic
 import Foundation
 
 /// Dependency graph node.
-class GraphNode: Equatable {
+class GraphNode: Equatable, Hashable {
     /// Node path.
     let path: AbsolutePath
 
@@ -15,6 +15,10 @@ class GraphNode: Equatable {
 
     static func == (lhs: GraphNode, rhs: GraphNode) -> Bool {
         return lhs.path == rhs.path
+    }
+
+    var hashValue: Int {
+        return path.hashValue
     }
 }
 
@@ -42,6 +46,10 @@ class TargetNode: GraphNode {
         self.target = target
         self.dependencies = dependencies
         super.init(path: project.path)
+    }
+
+    override var hashValue: Int {
+        return path.hashValue ^ target.name.hashValue
     }
 
     static func read(name: String, path: AbsolutePath, context: GraphLoaderContexting) throws -> TargetNode {